refactor: move InfoButtonFragment
