+++ app/src/main/res/layout/fragment_info_button.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.InfoButtonFragment"
+    tools:context=".ui.infobutton.InfoButtonFragment"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
     >
